speech_denoising/requirements.txt
# Speech Denoising Project - Requirements
# Python 3.8+ recommended

# Deep Learning Framework
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
scipy>=1.10.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Metrics
pesq>=0.0.4
pystoi>=0.3.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
hydra-core>=1.3.0

# For downloading dataset
requests>=2.31.0
